refactor(subprojects): subprojects nested inside projects, needs project id to retrieve subprojects
