+++ pages/projects/[id]/subprojects/[idSubproject].js
@@ -1,11 +1,11 @@
-import { useMockFetch } from "../../hooks/useFetch"
+import { useMockFetch } from "../../../../hooks/useFetch"
 import Head from "next/head"
-import Header from "../../components/header/Header"
-import { useUser } from "../../hooks/useUser"
+import Header from "../../../../components/header/Header"
+import { useUser } from "../../../../hooks/useUser"
 import { useEffect } from "react"
 import { useRouter } from "next/router"
-import PageSpinner from "../../components/PageSpinner"
-import SubprojectInfo from "../../components/subprojects/SubprojectInfo"
+import PageSpinner from "../../../../components/PageSpinner"
+import SubprojectInfo from "../../../../components/subprojects/SubprojectInfo"
 
 export default function Home() {
   const router = useRouter()
